Drafts of ExchangeActor and SettlementActor.
